Creates base test class for pypln tasks
